+++ index.js
@@ -7,6 +7,8 @@ module.exports = {
   verifyAndDecrypt: newebpay.verifyAndDecrypt,
   createQueryData: newebpay.createQueryData,
   queryTradeInfo: newebpay.queryTradeInfo,
+  closeTrade: newebpay.closeTrade,
+  cancelAuth: newebpay.cancelAuth,
   encryptTradeInfo: newebpay.encryptTradeInfo,
   decryptTradeInfo: newebpay.decryptTradeInfo,
   createTradeSha: newebpay.createTradeSha,